updated docker scripts

+++ docker-down-up.sh
@@ -2,4 +2,6 @@
 # -----------------------------------------------------------------------------
 # Brings all containers down and up again
 # -----------------------------------------------------------------------------
+
+: "${PKDB_DOCKER_COMPOSE_YAML:?The PKDB environment variable must be exported.}"
 docker-compose -f $PKDB_DOCKER_COMPOSE_YAML down && docker-compose -f $PKDB_DOCKER_COMPOSE_YAML up --detach
